将compressed_x改成float

+++ datatype.hpp
@@ -13,7 +13,7 @@
 typedef double original_data;
 typedef double x_value;
 typedef unsigned char index_value;
-typedef unsigned int compressed_x;
+typedef float compressed_x;
 typedef float original_data_write;
